5.1. Add board controller and move board rendering logics to it
